+++ packages/react-ui/components/Paging/Paging.styles.ts
@@ -12,11 +12,23 @@ export const styles = memoizeStyle({
     `;
   },
 
+  pagingDisabled() {
+    return css`
+      pointer-events: none;
+    `;
+  },
+
   dots(t: Theme) {
     return css`
       color: ${t.pagingDotsColor};
       display: inline-block;
       padding: ${t.pagingDotsPadding};
+    `;
+  },
+
+  dotsDisabled(t: Theme) {
+    return css`
+      color: ${t.textColorDisabled};
     `;
   },
 
@@ -51,7 +63,7 @@ export const styles = memoizeStyle({
     `;
   },
 
-  disabled(t: Theme) {
+  forwardLinkDisabled(t: Theme) {
     return css`
       color: ${t.pagingForwardLinkDisabledColor};
       cursor: default;
@@ -86,11 +98,24 @@ export const styles = memoizeStyle({
     `;
   },
 
-  active(t: Theme) {
+  pageLinkDisabled(t: Theme) {
+    return css`
+      color: ${t.linkDisabledColor};
+    `;
+  },
+
+  pageLinkCurrent(t: Theme) {
     return css`
       cursor: default;
       background: ${t.pagingPageLinkActiveBg} !important; // override hover styles
       color: ${t.pagingPageLinkActiveColor};
+    `;
+  },
+
+  pageLinkCurrentDisabled(t: Theme) {
+    return css`
+      background: ${t.pagingPageLinkDisabledActiveBg} !important;
+      color: ${t.linkDisabledColor};
     `;
   },