+++ packages/react-ui/components/Switcher/Switcher.tsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import { isKeyArrowHorizontal, isKeyArrowLeft, isKeyEnter } from '../../lib/events/keyboard/identifiers';
 import { Group } from '../Group';
-import { Button, ButtonSize } from '../Button';
+import { Button, ButtonProps, ButtonSize } from '../Button';
 import { Nullable } from '../../typings/utility-types';
 import { ThemeContext } from '../../lib/theming/ThemeContext';
 import { Theme } from '../../lib/theming/Theme';
@@ -23,7 +23,7 @@ export const SwitcherDataTids = {
 
 export interface SwitcherProps extends CommonProps {
   /**
-   * Список строк или список элементов типа `{ label: string, value: string }`
+   * Список строк или список элементов типа `{ label: string, value: string, buttonProps?: Partial<ButtonProps> }`
    */
   items: SwitcherItems[];
 
@@ -39,6 +39,17 @@ export interface SwitcherProps extends CommonProps {
   size?: SwitcherSize;
 
   disabled?: boolean;
+
+  /**
+   * Функция для отрисовки элемента. Аргументы — `label`,
+   * `value`, `buttonProps`, `renderDefault`
+   */
+  renderItem?: (
+    label: string,
+    value: string,
+    buttonProps: ButtonProps,
+    renderDefault: () => React.ReactNode,
+  ) => React.ReactNode;
 }
 
 export interface SwitcherState {
@@ -48,6 +59,7 @@ export interface SwitcherState {
 interface SwitcherItem {
   label: string;
   value: string;
+  buttonProps?: Partial<ButtonProps>;
 }
 
 @rootNode
@@ -69,6 +81,7 @@ export class Switcher extends React.Component<SwitcherProps, SwitcherState> {
     caption: PropTypes.string,
     value: PropTypes.string,
     onValueChange: PropTypes.func,
+    renderItem: PropTypes.func,
   };
 
   public state: SwitcherState = {
@@ -195,25 +208,36 @@ export class Switcher extends React.Component<SwitcherProps, SwitcherState> {
   };
 
   private _renderItems = () => {
-    return this.props.items.map((item, i) => {
-      const { label, value } = this._extractPropsFromItem(item);
-      const buttonProps = {
-        checked: this.props.value === value,
+    const { items, value, size, disabled, renderItem } = this.props;
+    return items.map((item, i) => {
+      const { label, value: itemValue, buttonProps: customButtonProps } = this._extractPropsFromItem(item);
+      const commonButtonProps = {
+        checked: value === itemValue,
         visuallyFocused: this.state.focusedIndex === i,
         onClick: () => {
-          this.selectItem(value);
+          this.selectItem(itemValue);
         },
         disableFocus: true,
-        size: this.props.size,
-        disabled: this.props.disabled,
+        size,
+        disabled,
       };
-      return (
-        <Button key={value} {...buttonProps}>
-          {label}
-        </Button>
-      );
+
+      const buttonProps = {
+        ...commonButtonProps,
+        ...customButtonProps,
+      };
+
+      const renderDefault = () => this.renderDefaultItem(label, itemValue, buttonProps);
+
+      return renderItem ? renderItem(label, itemValue, buttonProps, renderDefault) : renderDefault();
     });
   };
+
+  private renderDefaultItem = (label: string, value: string, buttonProps: ButtonProps) => (
+    <Button key={value} {...buttonProps}>
+      {label}
+    </Button>
+  );
 
   private getLabelSizeClassName = (): string => {
     switch (this.props.size) {