+++ packages/react-ui/components/Switcher/__stories__/Switcher.stories.tsx
@@ -1,26 +1,28 @@
 import React from 'react';
 
 import { Story } from '../../../typings/stories';
-import { Switcher } from '../Switcher';
+import { Switcher, SwitcherProps } from '../Switcher';
 import { Gapped } from '../../Gapped';
+import { Hint } from '../../Hint';
+import { Tooltip } from '../../Tooltip';
+import { ButtonProps } from '../../Button';
 
-interface ComponentProps {
-  items: string[];
-  error?: boolean;
+interface ComponentState {
+  value: string;
 }
-class Component extends React.Component<ComponentProps> {
-  public state = {
-    value: '',
-  };
+
+class Component extends React.Component<SwitcherProps, ComponentState> {
+  constructor(props: SwitcherProps) {
+    super(props);
+    this.state = {
+      value: props.value || '',
+    };
+  }
 
   public render() {
+    const { value, ...rest } = this.props;
     return (
-      <Switcher
-        value={this.state.value}
-        onValueChange={this.handleChange}
-        caption={'Label for Switcher'}
-        {...this.props}
-      />
+      <Switcher value={this.state.value} onValueChange={this.handleChange} caption={'Label for Switcher'} {...rest} />
     );
   }
 
@@ -75,4 +77,71 @@ export const Disabled = () => {
 Disabled.storyName = 'disabled';
 Disabled.parameters = {
   creevey: { skip: [{ in: ['chrome', 'chrome8px', 'chromeFlat8px', 'chromeDark'] }] },
+};
+
+const items: Array<{ label: string; value: string; buttonProps: Partial<ButtonProps> }> = [
+  {
+    label: 'One',
+    value: '111',
+    buttonProps: {
+      'data-tid': '1-1-1',
+      use: 'primary',
+    },
+  },
+  {
+    label: 'Two',
+    value: '222',
+    buttonProps: {
+      'data-tid': '2-2-2',
+      disabled: true,
+    },
+  },
+  {
+    label: 'Three',
+    value: '333',
+    buttonProps: {
+      'data-tid': '3-3-3',
+      use: 'danger',
+    },
+  },
+  {
+    label: 'four',
+    value: '444',
+    buttonProps: {
+      disabled: true,
+    },
+  },
+];
+
+const renderItem = (label: string, value: string, buttonProps: ButtonProps, renderDefault: () => React.ReactNode) => {
+  if (value === '111') {
+    return (
+      <Hint text="Текст Хинта" opened manual>
+        {renderDefault()}
+      </Hint>
+    );
+  }
+  if (value === '333') {
+    return (
+      <Tooltip pos="bottom center" trigger="opened" render={() => '⚠️ Лучше не трогай...'}>
+        {renderDefault()}
+      </Tooltip>
+    );
+  }
+  return renderDefault();
+};
+
+export const WithCustomRenderItems: Story = () => {
+  return (
+    <div style={{ padding: '65px 20px' }}>
+      <Component items={items} renderItem={renderItem} />
+    </div>
+  );
+};
+
+WithCustomRenderItems.storyName = 'with custom render item';
+WithCustomRenderItems.parameters = {
+  creevey: {
+    skip: { in: /^(?!\bchrome\b)/ },
+  },
 };